# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type CurrentUserRole = typeof CurrentUserRole[keyof typeof CurrentUserRole];


export const CurrentUserRole = {
  viewer: 'viewer',
  editor: 'editor',
} as const;

export interface CurrentUser {
  entraOid: string;
  email: string;
  displayName: string;
  role: CurrentUserRole;
}

export interface ErrorResponse {
  message: string;
}

export type ProductionOrderStatus = typeof ProductionOrderStatus[keyof typeof ProductionOrderStatus];


export const ProductionOrderStatus = {
  STARTED: 'STARTED',
  RELEASED: 'RELEASED',
} as const;

export type ProductionOrderLinkageStatus = typeof ProductionOrderLinkageStatus[keyof typeof ProductionOrderLinkageStatus];


export const ProductionOrderLinkageStatus = {
  Linked: 'Linked',
  Unlinked: 'Unlinked',
  Partial: 'Partial',
} as const;

/**
 * Deterministic reconciliation of the sales header, item line, and production demand lot.
 */
export interface SalesOrderLinkageAudit {
  salesOrder: string;
  /** Source data-area/company identifier. */
  company: string;
  itemNumber: string;
  inventoryLotDemandId: string;
  salesHeaderMatched: boolean;
  salesLineMatched: boolean;
  productionDemandMatched: boolean;
}

export interface SalesOrderLine {
  line: string;
  item: string;
  description: string;
  configuration: string;
  configName: string;
  /** @nullable */
  qty: number | null;
  unit: string;
  reference: string;
  status: string;
}

export interface ProductionOrder {
  id: string;
  workOrder: string;
  salesOrder: string;
  customer: string;
  customerPO: string;
  itemNumber: string;
  description: string;
  SC1: string;
  SC2: string;
  SC3: string;
  buildStartDate: string;
  buildEndDate: string;
  salesShipDate: string;
  createdDate: string;
  requestedDate: string;
  status: ProductionOrderStatus;
  currentStage: string;
  resource: string;
  rawResource: string;
  productionGroup: string;
  /** Full D365FO production-group name from the cost production group staging lookup. */
  productionGroupName: string;
  destination: string;
  country: string;
  salesperson: string;
  customerReference: string;
  exceptionFlags: string[];
  linkageStatus: ProductionOrderLinkageStatus;
  linkageAudit: SalesOrderLinkageAudit;
  /** All sales order lines on the order's sales order, sourced from the D365FO sales line staging table. */
  salesOrderLines: SalesOrderLine[];
}

export type FutureDemandStatus = typeof FutureDemandStatus[keyof typeof FutureDemandStatus];


export const FutureDemandStatus = {
  NO_WORK_ORDER: 'NO WORK ORDER',
} as const;

export interface FutureDemand {
  id: string;
  salesOrder: string;
  customer: string;
  customerPO: string;
  itemNumber: string;
  description: string;
  SC1: string;
  SC3: string;
  requestedDate: string;
  salesShipDate: string;
  /** @nullable */
  qty?: number | null;
  status: FutureDemandStatus;
}

export interface DataQuality {
  machineRows: number;
  validWorkOrders: number;
  duplicateWorkOrders: number;
  /** Compatibility count of orders with a populated sales-order number. */
  linkedOrders: number;
  /** Orders with a non-empty sales-order number. */
  populatedSalesOrders: number;
  /** Orders with all required sales header item-line company and inventory-lot demand identifiers matched. */
  validatedSalesOrderLinks: number;
  /** Orders with a sales-order number that fail one or more linkage checks. */
  partialLinkages: number;
  unlinkedOrders: number;
  futureDemand: number;
  futureDemandAlreadyLinked: number;
  invalidSC3: number;
  invalidDates: number;
}

export interface ProductionPriorityResponse {
  orders: ProductionOrder[];
  demand: FutureDemand[];
  dataQuality: DataQuality;
  dataAsOf: string;
}

export type BeginMicrosoftLoginParams = {
returnTo?: string;
};

export type CompleteMicrosoftLoginParams = {
code: string;
state: string;
};

export type GetProductionPriorityParams = {
/**
 * Bypass the short server-side read cache.
 */
refresh?: boolean;
};