# lib/api-zod/src/generated/types/currentUser.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { CurrentUserRole } from './currentUserRole';

export interface CurrentUser {
  entraOid: string;
  email: string;
  displayName: string;
  role: CurrentUserRole;
}