# lib/api-zod/src/generated/types/dataQuality.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface DataQuality {
  machineRows: number;
  validWorkOrders: number;
  duplicateWorkOrders: number;
  /** Compatibility count of orders with a populated sales-order number. */
  linkedOrders: number;
  /** Orders with a non-empty sales-order number. */
  populatedSalesOrders: number;
  /** Orders with all required sales header item-line company and inventory-lot demand identifiers matched. */
  validatedSalesOrderLinks: number;
  /** Orders with a sales-order number that fail one or more linkage checks. */
  partialLinkages: number;
  unlinkedOrders: number;
  futureDemand: number;
  futureDemandAlreadyLinked: number;
  invalidSC3: number;
  invalidDates: number;
}