# lib/api-zod/src/generated/types/futureDemand.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { FutureDemandStatus } from './futureDemandStatus';

export interface FutureDemand {
  id: string;
  salesOrder: string;
  customer: string;
  customerPO: string;
  itemNumber: string;
  description: string;
  SC1: string;
  SC3: string;
  requestedDate: string;
  salesShipDate: string;
  /** @nullable */
  qty?: number | null;
  status: FutureDemandStatus;
}