# lib/api-zod/src/generated/types/productionOrder.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ProductionOrderLinkageStatus } from './productionOrderLinkageStatus';
import type { ProductionOrderStatus } from './productionOrderStatus';
import type { SalesOrderLine } from './salesOrderLine';
import type { SalesOrderLinkageAudit } from './salesOrderLinkageAudit';

export interface ProductionOrder {
  id: string;
  workOrder: string;
  salesOrder: string;
  customer: string;
  customerPO: string;
  itemNumber: string;
  description: string;
  SC1: string;
  SC2: string;
  SC3: string;
  buildStartDate: string;
  buildEndDate: string;
  salesShipDate: string;
  createdDate: string;
  requestedDate: string;
  status: ProductionOrderStatus;
  currentStage: string;
  resource: string;
  rawResource: string;
  productionGroup: string;
  /** Full D365FO production-group name from the cost production group staging lookup. */
  productionGroupName: string;
  destination: string;
  country: string;
  salesperson: string;
  customerReference: string;
  exceptionFlags: string[];
  linkageStatus: ProductionOrderLinkageStatus;
  linkageAudit: SalesOrderLinkageAudit;
  /** All sales order lines on the order's sales order, sourced from the D365FO sales line staging table. */
  salesOrderLines: SalesOrderLine[];
}