# lib/api-zod/src/generated/types/productionPriorityResponse.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { DataQuality } from './dataQuality';
import type { FutureDemand } from './futureDemand';
import type { ProductionOrder } from './productionOrder';

export interface ProductionPriorityResponse {
  orders: ProductionOrder[];
  demand: FutureDemand[];
  dataQuality: DataQuality;
  dataAsOf: string;
}