# lib/api-zod/src/generated/types/salesOrderLine.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface SalesOrderLine {
  line: string;
  item: string;
  description: string;
  configuration: string;
  configName: string;
  /** @nullable */
  qty: number | null;
  unit: string;
  reference: string;
  status: string;
}