# lib/api-zod/src/generated/types/salesOrderLinkageAudit.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

/**
 * Deterministic reconciliation of the sales header, item line, and production demand lot.
 */
export interface SalesOrderLinkageAudit {
  salesOrder: string;
  /** Source data-area/company identifier. */
  company: string;
  itemNumber: string;
  inventoryLotDemandId: string;
  salesHeaderMatched: boolean;
  salesLineMatched: boolean;
  productionDemandMatched: boolean;
}